refactor: move form validation error and pass all errors
