+++ packages/common/src/errors/FormValidationError.ts
@@ -4,7 +4,7 @@ export class FormValidationError extends Error {
   public errors: FormValidationErrors;
 
   constructor(errors: FormValidationErrors) {
-    super(Object.values(errors)[0][0]);
+    super(Object.values(errors).flat().join(';'));
 
     this.errors = errors;
   }
